requirements.txt
transformers==4.38.2
python-dotenv==1.0.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
requests==2.31.0
ipython==8.22.2